moved fixtures to a central place
